pyproject.toml
[project]
name = "validate-pathology-diagnosis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pydantic-autocli>=0.3.2",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vpd"]

[project.scripts]
vpd = "vpd.cli:main"
